Combined all database templates to one single file
Changed all possible elements from each page to a template call.  Refer to app.py for variables passed in,
Since iterating through keys is undefined in terms of sorting, we have to manually define the columns and
pass it in through flask.  I also currently have angular loading at the top of the page, which may be a
cause for slow page loading.